requirements.txt
requests
psycopg2-binary
dataclasses
python-dotenv
fastapi
inquirerPy
datetime